feat(agents): split manage, sessions, and detail nav views
Move agent cards and session lists into agent_ui, promote AgentDetail to
its own workspace main route, and add bg_card_dim for quieter list rows.
