+++ crates/zedra/src/workspace_state.rs
@@ -5,6 +5,7 @@ use std::sync::{Mutex, OnceLock};
 use tracing::*;
 use zedra_rpc::proto::HostInfoSnapshot;
 
+use zedra_rpc::proto::ManagedAgentKind;
 use zedra_session::*;
 
 use crate::platform_bridge;
@@ -41,6 +42,9 @@ pub enum WorkspaceMainView {
     },
     AgentSessions,
     AgentManage,
+    AgentDetail {
+        kind: ManagedAgentKind,
+    },
     NoActiveTerminal,
 }
